+++ modules/task_2/ndanilov_topologies/include/host.h
@@ -4,7 +4,7 @@
 
 #include <string>
 
-#include "packet.h"
+#include "include/packet.h"
 
 class host {
  public:
